Print out raw versions of exemplar tweets
The cleaned versions of exemplar tweets are no longer printed out; instead, the raw versions of those same tweets are.
In order to make the 1-to-1 comparison between the list of tweet embedding vectors and the list of raw tweets, both lists have to be the same size. To ensure this, `processing.average_vectors()` was changed to append the zero vector when reading a tweet made empty by the cleaning process, instead of omitting it.